+++ src/components/ExpandCollapse/__stories__/expandCollapse.stories.js
@@ -5,16 +5,23 @@ import Icon from "../../Icon/Icon";
 import Robot from "../../Icon/Icons/components/Robot";
 import "./expandCollapse.stories.scss";
 
+function HeaderComponent() {
+  return <span>I can be anything</span>;
+}
+
+function H1Component() {
+  return <h1>Any component you want</h1>;
+}
 export const Sandbox = () => {
-  const width =  300;
-  const height =  200;
+  const width = 300;
+  const height = 200;
 
   return (
     <div style={{ width, height }}>
-      <ExpandCollapse headerComponentRenderer={() => <span>I can be anything</span>}>
+      <ExpandCollapse headerComponentRenderer={HeaderComponent}>
         <h2>insert any component you want</h2>
         <p>here is a robot for you</p>
-        <Icon iconType={Icon.type.SVG} icon={Robot} iconSize={"52px"} tabindex="-1" clickable={true} />
+        <Icon iconType={Icon.type.SVG} icon={Robot} iconSize="52px" tabindex="-1" clickable={true} />
       </ExpandCollapse>
     </div>
   );
@@ -25,9 +32,9 @@ export const OpenByDefault = () => (
     <ExpandCollapse
       className="expandCollapse"
       defaultOpenState={true}
-      headerComponentRenderer={() => <h1>Any component you want</h1>}
+      headerComponentRenderer={H1Component}
     >
-      <Icon iconType={Icon.type.SVG} icon={Robot} iconSize={"52px"} tabindex="-1" clickable={true} />
+      <Icon iconType={Icon.type.SVG} icon={Robot} iconSize="52px" tabindex="-1" clickable={true} />
     </ExpandCollapse>
   </div>
 );