+++ src/components/Toast/__tests__/toast-tests.jest.js
@@ -1,7 +1,6 @@
 import React from "react";
-import { fireEvent, render, cleanup } from "@testing-library/react";
+import { fireEvent, render } from "@testing-library/react";
 import { act } from "@testing-library/react-hooks";
-import ToastButton from "../ToastButton/ToastButton";
 import Toast from "../Toast";
 
 const renderComponent = props => {