+++ src/hooks/__tests__/useDebounceEvent.jest.js
@@ -1,4 +1,3 @@
-import React from "react";
 import { renderHook, cleanup, act } from "@testing-library/react-hooks";
 import useDebounceEvent from "../useDebounceEvent";
 
@@ -15,8 +14,7 @@ describe("useDebounceEvent", () => {
         delay,
         initialStateValue,
         onChange: onChangeCallbackStub
-      })
-    );
+      }));
   });
 
   afterEach(() => {
@@ -59,8 +57,7 @@ describe("useDebounceEvent", () => {
           trim: true,
           onChange: onChangeCallbackStub,
           initialStateValue: ""
-        })
-      );
+        }));
 
       const { onEventChanged } = hookRes.result.current;
       const newInputValue = "value     ";
@@ -102,8 +99,7 @@ describe("useDebounceEvent", () => {
           delay: additionalDelay,
           initialStateValue,
           onChange: onChangeCallbackStub
-        })
-      );
+        }));
     });
 
     afterEach(() => {