+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 1.3.1 (2022-01-13)
+
+#### Bug Fixes
+* [#457](https://github.com/mondaycom/monday-ui-react-core/pull/457) fix: Loader issues ([@orrgottlieb](https://github.com/orrgottlieb))
+
+## 1.3.0 (2022-01-13)
+
+#### New Features
+* [#443](https://github.com/mondaycom/monday-ui-react-core/pull/443) Adding VirtualizedGrid component ([@or-nuri-monday](https://github.com/or-nuri-monday))
+
+#### Dependency Upgrades
+* [#455](https://github.com/mondaycom/monday-ui-react-core/pull/455) Fix/eslint ([@orrgottlieb](https://github.com/orrgottlieb))
+
+## 1.2.6 (2022-01-12)
+
+#### New Icons
+* [#454](https://github.com/mondaycom/monday-ui-react-core/pull/454) Upgrade icons ([@github-actions[bot]](https://github.com/apps/github-actions))
+
+## 1.2.5 (2022-01-12)
+
+#### Bug Fixes
+* [#453](https://github.com/mondaycom/monday-ui-react-core/pull/453) Fix icon button export ([@hadasfa](https://github.com/hadasfa))
+* [#452](https://github.com/mondaycom/monday-ui-react-core/pull/452) Fix dialog hideTriggerIgnoreClass prop type to accept array ([@MosheZemah](https://github.com/MosheZemah))
+
 ## 1.2.4 (2022-01-11)
 
 #### Bug Fixes